fix: start job runs in workflows controller
